refactor: extract useFileThumbnail
